requirements.txt
flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.10.3
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai==0.3.2
pymongo==4.6.0
requests==2.31.0
python-docx==1.1.0